+++ flink_jobs_v2/stream_status/pom.xml
@@ -63,6 +63,12 @@ language governing permissions and limitations under the License. -->
     handling. This approach is preferred and leads to much cleaner jar files. -->
 
     <dependencies>
+         <dependency>
+            <groupId>ams.connector</groupId>
+            <artifactId>ams-connector</artifactId>
+            <version>2.0.0</version>
+            <type>jar</type>
+        </dependency>
         <dependency>
             <groupId>api.resource.manager</groupId>
             <artifactId>ApiResourceManager</artifactId>
